requirements.txt
asgiref==3.3.4
attrs==20.3.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
importlib-metadata==4.0.1
jsonschema==3.2.0
psycopg2==2.8.6
PyJWT==2.1.0
pyrsistent==0.17.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.1
